feat:added-better-types
