+++ src/lib.rs
@@ -6,6 +6,7 @@ pub mod order_gateway;
 pub mod protocol;
 mod routing;
 pub mod types;
+pub mod ws_utils;
 
 pub use client::ArchitectX;
